+++ src/chains/optimism/fees.ts
@@ -1,4 +1,4 @@
-import type { ChainFees } from '../../chains/types.js'
+import type { ChainFees } from '../types.js'
 
 export const feesOptimism = {
   defaultPriorityFee: 1_000_000n, // 0.001 gwei
