+++ src/chains/optimism/formatters.ts
@@ -1,4 +1,3 @@
-import { type ChainFormatters } from '../../chains/types.js'
 import type { Hash } from '../../types/misc.js'
 import { type RpcTransaction } from '../../types/rpc.js'
 import { hexToBigInt } from '../../utils/encoding/fromHex.js'
@@ -8,6 +7,7 @@ import {
   formatTransaction,
 } from '../../utils/formatters/transaction.js'
 import { defineTransactionReceipt } from '../../utils/formatters/transactionReceipt.js'
+import { type ChainFormatters } from '../types.js'
 import type {
   OptimismBlockOverrides,
   OptimismRpcBlockOverrides,
